Implemented Add Items via Admin
I implemented being able to create new items as an admin. The SellerHomePage frame looks ugly but it is functional and shows the appropriate informtion so far. It does not update when you input new item - will do that later. Also tinkered with the login page to direct you to the seller home page just to see what it looks like. feel free to change it back. We also need to figure out how to determine if a user is an admin or a customer